+++ tests/e2e/specs/identity/identity.empty-file.all.e2e.ts
@@ -1,9 +1,10 @@
-import { runIdentityTestSuite } from '../../helpers/identity.shared'
+import { runIdentityTests } from './identity.shared'
 
-runIdentityTestSuite({
-  suiteName: 'Test suite: Empty file with setting All (1)',
-  lifecycleSetting: 'All',
-  fixtureFile: '/tests/fixtures/identity/empty-file.md',
-  expectedOutput: '',
-  revertFile: 'empty-file.md',
+describe('Test suite: Empty file with setting All (1)', () => {
+  runIdentityTests({
+    lifecycleSetting: 'All',
+    fixtureFile: '/tests/fixtures/identity/empty-file.md',
+    expectedOutput: '',
+    revertFile: 'empty-file.md',
+  })
 })
