+++ tests/e2e/specs/identity/identity.empty-file.cell.e2e.ts
@@ -1,9 +1,10 @@
-import { runIdentityTestSuite } from '../../helpers/identity.shared'
+import { runIdentityTests } from './identity.shared'
 
-runIdentityTestSuite({
-  suiteName: 'Test suite: Empty file with setting Cell (3)',
-  lifecycleSetting: 'Cell',
-  fixtureFile: '/tests/fixtures/identity/empty-file.md',
-  expectedOutput: '',
-  revertFile: 'empty-file.md',
+describe('Test suite: Empty file with setting Cell only (3)', () => {
+  runIdentityTests({
+    lifecycleSetting: 'Cell',
+    fixtureFile: '/tests/fixtures/identity/empty-file.md',
+    expectedOutput: '',
+    revertFile: 'empty-file.md',
+  })
 })
