+++ tests/e2e/specs/identity/identity.empty-file.document.e2e.ts
@@ -1,9 +1,10 @@
-import { runIdentityTestSuite } from '../../helpers/identity.shared'
+import { runIdentityTests } from './identity.shared'
 
-runIdentityTestSuite({
-  suiteName: 'Test suite: Empty file with setting Document (2)',
-  lifecycleSetting: 'Doc',
-  fixtureFile: '/tests/fixtures/identity/empty-file.md',
-  expectedOutput: '',
-  revertFile: 'empty-file.md',
+describe('Test suite: Empty file with setting Document (2)', () => {
+  runIdentityTests({
+    lifecycleSetting: 'Doc',
+    fixtureFile: '/tests/fixtures/identity/empty-file.md',
+    expectedOutput: '',
+    revertFile: 'empty-file.md',
+  })
 })
