+++ tests/e2e/specs/identity/identity.empty-file.none.e2e.ts
@@ -1,10 +1,11 @@
-import { runIdentityTestSuite } from '../../helpers/identity.shared'
+import { runIdentityTests } from './identity.shared'
 
-runIdentityTestSuite({
-  suiteName: 'Test suite: Empty file with setting None (0)',
-  lifecycleSetting: 'None',
-  fixtureFile: '/tests/fixtures/identity/empty-file.md',
-  expectedOutput: '',
-  revertFile: 'empty-file.md',
-  assertOptions: { strict: true },
+describe('Test suite: Empty file with setting None (0)', () => {
+  runIdentityTests({
+    lifecycleSetting: 'None',
+    fixtureFile: '/tests/fixtures/identity/empty-file.md',
+    expectedOutput: '',
+    revertFile: 'empty-file.md',
+    assertOptions: { strict: true },
+  })
 })
