+++ tests/e2e/specs/identity/identity.shebang.all.e2e.ts
@@ -1,11 +1,11 @@
-import { runIdentityTestSuite } from '../../helpers/identity.shared'
+import { runIdentityTests } from './identity.shared'
 
-runIdentityTestSuite({
-  suiteName: 'Test suite: Shebang with setting All (1)',
-  lifecycleSetting: 'All',
-  fixtureFile: '/tests/fixtures/identity/shebang.md',
-  cellSelector: 'console.log("Scenario 1: Run scripts via Shebang!")',
-  expectedOutput: `---
+describe('Test suite: Shebang with setting All (1)', () => {
+  runIdentityTests({
+    lifecycleSetting: 'All',
+    fixtureFile: '/tests/fixtures/identity/shebang.md',
+    cellSelector: 'console.log("Scenario 1: Run scripts via Shebang!")',
+    expectedOutput: `---
       runme:
         id: 01HEXJ9KWG7BYSFYCNKSRE4JZR
         version: v3
@@ -16,11 +16,12 @@ runIdentityTestSuite({
 
       ## Scenario 1
 
-      \`\`\`js {"id":"01HEXJ9KWG7BYSFYCNKVF0VWR6","name":"foo"}
+      \`\`\`js {"name":"foo","id":"01HEXJ9KWG7BYSFYCNKVF0VWR6"}
       console.log("Scenario 1: Run scripts via Shebang!")
 
       \`\`\`
 
       `,
-  revertFile: 'shebang.md',
+    revertFile: 'shebang.md',
+  })
 })
